Move components into components folder. Rename components to better reflect domain modeling.
