+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import '../styles/App.css'
-import StopWatch from './StopWatch'
+import StopWatch from './components/StopWatch'
 import '../styles/StopWatch.css'
 
 export default function App() {
