--- styles/StopWatchButton.css
@@ -1,52 +1,0 @@
-.controls button {
-  margin: 15px;
-  font-size: large;
-  border: none;
-  /* background-color: #63e3e1; */
-  font-family: 'Roboto', sans-serif;
-  border-radius: 10px;
-  cursor: pointer;
-  width: 60px;
-}
-
-.lap-container {
-  height: auto;
-  border: 2px solid #939d9b;
-  max-height: 200px;
-  min-height: 200px;
-  overflow: auto;
-  font-family: 'Roboto', sans-serif;
-  border-radius: 10px;
-  font-size: large;
-  background-color: rgb(247, 247, 244);
-  color: black;
-  overflow: overlay;
-}
-
-::-webkit-scrollbar-track {
-  background-color: transparent;
-  margin-block: .5em;
-  border-radius: 10px;
-}
-
-.lap {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  padding-left: 20px;
-}
-
-.start {
-  background-color: #53c24d;
-}
-
-.stop {
-  background-color: rgb(212, 61, 61);
-}
-
-.reset {
-  background-color: orange;
-}
-
-.lapButton {
-  background-color: #63e3e1;
-}
